refactor: extract tab rendering and layout from FolderTabbedPane (643→370 lines)
Extract 4 inner classes from FolderTabbedPane.java into two new files:

- FolderTabRenderer.java (216 lines): FolderTabTitleUI, JFolderTabTitle,
  FolderTabTitle — the tab rendering stack
- FolderTitlesPanel.java (176 lines): promoted from TitlesPanel inner class —
  tab layout and painting

Key changes:
- Constants TRAILING_TAB_PAD/OUTLINE_THICKNESS widened to package-private
- FolderTabTitle receives explicit SingleSelectListState param (replaces
  implicit outer-class reference)
- titlesPanel field type and createTitlesPanel() return type updated to
  FolderTitlesPanel
- PopupOperation, PopupButton, ScrollListener remain as inner classes

All 47 extraction tests pass. Maven compile succeeds.
